+++ src/config/directAds.js
@@ -1,0 +1,14 @@
+import LambdaClassesPoster from "../images/lambda-classes-projects-poster.png";
+
+export const DIRECT_ADS = {
+    lambdaClasses: {
+        id: "lambda-classes",
+        href: "https://www.lambda-classes.com/#apply",
+        image: LambdaClassesPoster,
+        alt: "Lambda Classes",
+        trackingEvent: "lc clickthrough",
+        trackingProps: { ad: "lambda-classes" },
+        target: "_blank",
+        rel: "noreferrer",
+    },
+};